Expanded info panel and created stats panel

+++ Assets/2 - Scripts/ScriptableObjects/Items/Equippables/Weapons/Guns/_TestGuns/TestGun.cs
@@ -12,7 +12,7 @@ public class TestGun : Item, IGun
     [SerializeField]
     private float ySpread = 0.0f;
     [SerializeField]
-    private float maxSlots = 0.0f;
+    private int maxSlots = 0;
 
     private void Awake()
     {
@@ -25,5 +25,5 @@ public class TestGun : Item, IGun
     public float RechargeTime { get => rechargeTime; set => rechargeTime = value; }
     public float XSpread { get => xSpread; set => xSpread = value; }
     public float YSpread { get => ySpread; set => ySpread = value; }
-    public float MaxSlots { get => maxSlots; set => maxSlots = value; }
+    public int MaxSlots { get => maxSlots; set => maxSlots = value; }
 }
