+++ Assets/2 - Scripts/ScriptableObjects/Items/Equippables/Weapons/_Interfaces/IWeapon.cs
@@ -4,5 +4,5 @@ public interface IWeapon : IEquippable
     public float RechargeTime { get; set; }
     public float XSpread { get; set; }
     public float YSpread { get; set; }
-    public float MaxSlots { get; set; }
+    public int MaxSlots { get; set; }
 }
